feat/SEOULDATA-51 refactor: fest 관련 파일 구조 변경
