+++ fest/src/main/java/com/seouldata/fest/FestApplication.java
@@ -2,7 +2,11 @@ package com.seouldata.fest;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
+@EnableScheduling
+@EnableFeignClients(basePackages = {""})
 @SpringBootApplication
 public class FestApplication {
