+++ fest/src/main/java/com/seouldata/fest/domain/fest/entity/Fest.java
@@ -1,4 +1,4 @@
-package com.seouldata.fest.domain.entity;
+package com.seouldata.fest.domain.fest.entity;
 
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
