+++ fest/src/main/java/com/seouldata/fest/domain/fest/repository/FestRepository.java
@@ -1,6 +1,6 @@
-package com.seouldata.fest.domain.repository;
+package com.seouldata.fest.domain.fest.repository;
 
-import com.seouldata.fest.domain.entity.Fest;
+import com.seouldata.fest.domain.fest.entity.Fest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
